% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{BetaTest}
\alias{BetaTest}
\title{BetaTest}
\usage{
BetaTest(output, outputRegpcond, alfa = 0.05)
}
\arguments{
\item{output}{Output object of \link{more} function when MLR model is used.}

\item{outputRegpcond}{Output object of \link{RegulationPerCondition} function.}

\item{alfa}{Significance level. By default, 0.05.}
}
\value{
Modified outputRegpcond table with zeros in the coefficients that did not present statistically significant differences.
}
\description{
\code{BetaTest} Function to test if the coefficients for each condition are statistically different from zero.
}
