% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostatcolors.R
\name{biostat_cols}
\alias{biostat_cols}
\title{Function to extract biostat colors as hex codes}
\usage{
biostat_cols(...)
}
\arguments{
\item{...}{Character names of biostat_colors}
}
\description{
Function to extract biostat colors as hex codes
}
