% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostatcolors.R
\name{biostat_pal}
\alias{biostat_pal}
\title{Interpolate a BioStatOmics Color Palette}
\usage{
biostat_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{A character string specifying the name of the desired palette from the `biostat_palettes`. Available options include: "main", "cblindfriendly", "sunshine", "hot", "warm" and "cold".}

\item{reverse}{A logical value indicating whether the colors in the selected palette should be reversed. Default is `FALSE`.}

\item{...}{Additional arguments to be passed to the `colorRampPalette` function from the `grDevices` package.}
}
\description{
The `biostat_pal` function offers a flexible way to interpolate colors from the BiostatOmics color palettes. This function provides an interface to generate a range of colors based on the selected BioStatOmics palette, allowing for enhanced customization in scientific visualizations.
}
\details{
The BioStatOmics color palettes, available in the BioStatOmicsColors package, have been specifically curated for scientific visualizations. The `biostat_pal` function leverages the `colorRampPalette` function from the `grDevices` package to interpolate between the colors of the chosen BioStatOmics palette. This interpolation capability ensures that users can generate a continuous range of colors, suitable for representing a wide variety of data types and scales. Whether visualizing continuous data gradients or categorical distinctions, the interpolated BiostatOmics palettes can provide clarity and aesthetic appeal to the visual representation.
}
\author{
Pedro Salguero Garcia, Maider Aguerralde Martin. Maintainer: magumar2@upv.edu.es
}
