% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostatcolors.R
\docType{data}
\name{biostat_palettes}
\alias{biostat_palettes}
\title{Biostatomics color palette}
\format{
An object of class \code{list} of length 8.
}
\usage{
biostat_palettes
}
\description{
Biostatomics color palette
}
\keyword{datasets}
