% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostatcolors.R
\name{colorbiostat}
\alias{colorbiostat}
\title{Retrieve Colors from BioStatOmics group's Palettes}
\usage{
colorbiostat(n, reverse = FALSE, palette = "main")
}
\arguments{
\item{n}{An integer specifying the number of colors to be extracted from the chosen palette.}

\item{reverse}{A logical value indicating whether the colors in the selected palette should be reversed (Default: `FALSE`).}

\item{palette}{A character string specifying the name of the desired palette from the `biostat_palettes`. Available options include: "main", "complete", "cblindfriendly", "sunshine", "hot", "warm", "grass" and "oficial" (Default: "main").}
}
\value{
A character vector of colors corresponding to the specified number and palette.
}
\description{
The `colorbiostat` function facilitates the extraction of a specified number of colors from the BioStatOmics group's curated color palettes. This function is designed to obtain a set of colors for their scientific visualizations.
}
\details{
BioStatOmics group's color palettes, available within the package, are tailored for scientific data visualization. The `colorbiostat` function is built upon these palettes, offering flexibility in color selection based on the user's requirements. It integrates with the `palette` argument to choose the color thematic.

It's essential to note that if the requested number of colors (`n`) is less than or equal to the size of the chosen palette, the function will directly extract the colors without interpolation. However, if `n` surpasses the palette size, interpolation is employed to generate the required colors.
}
\examples{
library(ggplot2)
data("iris")
colorSpecies <- colorbiostat(3, palette = "oficial")
plot(x = iris$Sepal.Length, y = iris$Sepal.Width, col = colorSpecies[iris$Species], pch = 16)

}
\author{
Pedro Salguero Garcia, Maider Aguerralde Martin. Maintainer: magumar2@upv.edu.es
}
