% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeGLM_function.R
\name{ComputeGLM}
\alias{ComputeGLM}
\title{Title}
\usage{
ComputeGLM(matrix.temp, alfa = 0.05, stepwise = "two.ways.backward",
  Res.df = 4, family = MASS::negative.binomial(theta = 10),
  epsilon = 1e-05, MT.adjust = "fdr", iter.max = 100)
}
\arguments{
\item{matrix.temp}{}

\item{alfa}{}

\item{stepwise}{}

\item{Res.df}{}

\item{family}{}

\item{epsilon}{}

\item{MT.adjust}{}

\item{iter.max}{}
}
\description{
Title
}
\keyword{internal}
