% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{differentialRegPlot}
\alias{differentialRegPlot}
\title{differentialRegPlot}
\usage{
differentialRegPlot(output, outputRegpcond)
}
\arguments{
\item{output}{MORE object obtained from applying \link{more} function.}

\item{outputRegpcond}{Object generated by the function \link{RegulationPerCondition} when applied to a \link{more} object.}
}
\value{
UpSet plot of the differential regulators across conditions and omics induced from more.
}
\description{
\code{differentialRegPlot} Function to be applied to more and RegulationPerCondition functions outputs. Creates an upset plot to analyze the differential regulators across conditions and omics.
}
