% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{FilterRegulationPerCondition}
\alias{FilterRegulationPerCondition}
\title{FilterRegulationPerCondition}
\usage{
FilterRegulationPerCondition(output, outputRegpcond, filterR2 = 0)
}
\arguments{
\item{output}{Output object of \link{more} function.}

\item{outputRegpcond}{Output object of \link{RegulationPerCondition} function.}

\item{filterR2}{Filters the results for the genes that showed a R2 above the one indicated. By default, 0.}
}
\value{
RegulationPerCondition matrix filtered to the genes with a R2 above the indicated.
}
\description{
\code{FilterRegulationPerCondition} Function to be applied to \link{more} and \link{RegulationPerCondition} function outputs.
}
