% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostatcolors.R
\name{getbiostatColors}
\alias{getbiostatColors}
\title{Retrieve BioStatOmics Main Color Set}
\usage{
getbiostatColors()
}
\value{
A list containing the primary colors from the BioStatOmics collection. Each color in the list is represented as a hexadecimal color value.
}
\description{
The `getbiostatColors` function provides access to a curated set of colors that are part of the BiostatOmicsColors package. These colors have been specifically chosen for their utility in scientific visualizations.
}
\details{
When using the `getbiostatColors` function, users can seamlessly integrate these colors into their R visualizations, benefiting from the expertise embedded in the BiostatOmics color selection.
}
\examples{
getbiostatColors()

}
\author{
Pedro Salguero Garcia, Maider Aguerralde Martin. Maintainer: magumar2@upv.edu.es
}
