% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biostatcolors.R
\name{getbiostatPalettes}
\alias{getbiostatPalettes}
\title{Return Biostatomics Color Palettes}
\usage{
getbiostatPalettes()
}
\value{
A list containing the various color palettes from the BioStatOmics collection. Each palette in the list is represented as a vector of color values.
}
\description{
The `getbiostatPalettes` function retrieves a collection of color palettes, specifically designed for scientific visualizations. These palettes are part of the biostat collection.
}
\details{
By using the `getbiostatPalettes` function, users can access these palettes and incorporate them into their visualizations, ensuring that their plots and graphs are both informative and visually appealing.
}
\examples{
getbiostatPalettes()

}
\author{
Pedro Salguero Garcia, Maider Aguerralde Martin. Maintainer: magumar2@upv.edu.es
}
