% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MORE_GLM.R
\name{GetGLM}
\alias{GetGLM}
\title{Genome-wide Generalized Linear Models}
\usage{
GetGLM(GeneExpression, associations, data.omics, edesign = NULL,
  cont.var = "Time", degree = 1, center = TRUE, scale = FALSE,
  Res.df = 5, epsilon = 1e-05, alfa = 0.05, MT.adjust = "none",
  family = negative.binomial(theta = 10), elasticnet = 0.5,
  stepwise = "backward", interactions.exp = TRUE,
  interactions.reg = 1, min.variation = 0, correlation = 0.9,
  min.obs = 10, omic.type = 0, mc.cores = 1)
}
\arguments{
\item{GeneExpression}{Data frame containing gene expression data with genes in rows and
experimental samples in columns. Row names must be the gene IDs.}

\item{associations}{List where each element correspond to a different omic data type (miRNAs,
transcription factors, methylation, etc.). The names of the list will be the omics. Each element
is a data frame with 2 columns (optionally 3) describing the potential interactions between genes
and regulators for that omic. First column must contain the genes (or features in
GeneExpression object), second column must contain the regulators, and an additional column can
be added to describe the type of interaction (e.g., for methylation, if a CpG site is located in
the promoter region of the gene, in the first exon, etc.).}

\item{data.omics}{List where each element correspond to a different omic data type (miRNAs,
transcription factors, methylation, etc.). The names of the list will be the omics. Each element
is a data matrix with omic regulators in rows and samples in columns.}

\item{edesign}{Data frame describing the experimental design. Rows must be the samples (columns
in GeneExpression) and columns must be the experimental variables to be included in the model
(e.g. time, treatment, etc.).}

\item{cont.var}{Name of the column in edesign that is to be considered as a continuous
explanatory variable in the regression model. NULL if edesign does not contain any continuous
variables. By default, "Time".}

\item{degree}{If cont.var is not NULL, non-linear (polynomial) relationships between the cont.var
and the gene expression must be studied. By default, degree = 1 and must be an integer.
A higher number will allow for in quadratic, cubic, etc. terms for cont.var in the model.}

\item{Res.df}{Number of degrees of freedom in the residuals. By default, 5. Increasing
Res.df will increase the power of the statistical model.}

\item{alfa}{Significance level. By default, 0.05.}

\item{MT.adjust}{Multiple testing correction method to be used within the stepwise variable
selection procedure. By default, "none". See the different options in ?\code{p.adjust}.}

\item{family}{Error distribution and link function to be used in the model (see ?\code{glm}
for more information). By default, negative.binomial(theta=10).}

\item{elasticnet}{ElasticNet mixing parameter. NULL = No ElasticNet variable selection. A number between 0 and 1 = ElasticNet is applied
with this number being the combination between Ridge and Lasso penalization (elasticnet=0 is the ridge penalty, elasticnet=1 is the lasso
penalty). The default is 0.5, which gives equal weight to ridge and lasso.}

\item{stepwise}{Stepwise variable selection method to be applied. It can be one of: "none", "backward"
(default), "forward", "two.ways.backward" or "two.ways.forward".}

\item{interactions.exp}{If TRUE (default), interactions among the experimental variables
are included in the model.}

\item{interactions.reg}{If TRUE (default), interactions between regulators and experimental
variables are included in the model.}

\item{min.variation}{For numerical regulators, the minimum change that a regulator must present across conditions
to keep it in the regression models. For binary regulators, if the proportion of the most repeated value equals or exceeds this value,
the regulator will be considered to have low variation and removed from the regression models.}

\item{omic.type}{0 = numerical variable (default), 1 = binary (categorical) variable. It should be a vector with length the number of omic types or a number (0 or 1) if all the omics are of the same type.}
}
\description{
\code{GetGLM} fits a regression model for all the genes in the data set to identify
the experimental variables and potential regulators that show a significant effect on
the expression of each gene.
}
