% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{globalreg_plot}
\alias{globalreg_plot}
\title{globalreg_plot}
\usage{
globalreg_plot(output_regincond, by_network = FALSE)
}
\arguments{
\item{output_regincond}{Output object of running \link{RegulationInCondition} function.}

\item{by_network}{If TRUE, information would be plotted on a network instead of a corrplot. By default, FALSE.}
}
\value{
Graphical visualization between global regulators and genes in a specific condition.
}
\description{
\code{globalreg_plot} Function to be applied to \link{RegulationInCondition} function output.
}
