% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{globalregPlot}
\alias{globalregPlot}
\title{globalregPlot}
\usage{
globalregPlot(outputRegincond, byNetwork = FALSE)
}
\arguments{
\item{outputRegincond}{Output object of running \link{RegulationInCondition} function.}

\item{byNetwork}{If TRUE, information would be plotted on a network instead of a corrplot. By default, FALSE.}
}
\value{
Graphical visualization between global regulators and target features in a specific condition.
}
\description{
\code{globalregPlot} Function to be applied to \link{RegulationInCondition} function output.
}
