% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{GSEA_more}
\alias{GSEA_more}
\title{GSEA_more}
\usage{
GSEA_more(
  output_globregincond,
  output_globregincond2 = NULL,
  annotation,
  alpha = 0.05,
  p.adjust.method = "fdr"
)
}
\arguments{
\item{output_globregincond}{Output object of running RegulationInCondition function}

\item{output_globregincond2}{Output object of running RegulationInCondition function for other group different to the previous. By default, NULL.}

\item{annotation}{Annotation matrix with genes in the first column, GO terms in the second}

\item{alpha}{The adjusted pvalue cutoff to consider}

\item{p.adjust.method}{One of holm, hochberg, hommel, bonferroni, BH, BY, fdr or none}
}
\value{
Plot of the network induced from more.
}
\description{
\code{GSEA_more} Function to be applied to RegulationInCondition function output.
}
