% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{gseaMORE}
\alias{gseaMORE}
\title{gseaMORE}
\usage{
gseaMORE(
  outputRegincond,
  outputRegincond2 = NULL,
  annotation,
  alpha = 0.05,
  p.adjust.method = "fdr"
)
}
\arguments{
\item{outputRegincond}{Output object of running RegulationInCondition function}

\item{outputRegincond2}{Object generated by the function \link{RegulationInCondition} for another group different from the previously considered. By default, NULL. If NULL, the analysis will be centered only in the study group, so the results will apply only to that group and will not mean the statistical differences between groups. If provided MORE computes and internal score comparing the number of significant regulators in both groups.}

\item{annotation}{Annotation matrix with target features in the first column, pathway ID in the second and pathway name in the third. Only necessary when a specific pathway has to be plotted. By default, NULL.}

\item{alpha}{p-value cutoff to consider. By default, 0.05.}

\item{p.adjust.method}{One of holm, hochberg, hommel, bonferroni, BH, BY, fdr or none}
}
\value{
Plot of the network induced from more.
}
\description{
\code{gseaMORE} Function to be applied to RegulationInCondition function output.
}
