% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxFunctions.R
\name{LowVariatFilter}
\alias{LowVariatFilter}
\title{Low variation filtering}
\usage{
LowVariatFilter(data, method, percVar, omic.type)
}
\arguments{
\item{data}{For computing variability. It could be mean values between replicates}

\item{method}{One of "sd","range", "IQrange" or "user"}

\item{percVar}{percentage of variation defined by the user}

\item{omic.type}{}
}
\value{

}
\description{
Low variation filtering
}
\keyword{internal}
