% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_message.R
\name{more_message}
\alias{more_message}
\title{Constructor for more messages definded in-line.}
\usage{
more_message(title, body)
}
\arguments{
\item{title}{the title of message}

\item{body}{the rest of the message}
}
\description{
This is a simple helper for defining messages with a title and body using
formatting from \code{{cli}}.
}
\details{
Messages can be any text, so you are not limited to using this function to
create them.
}
