% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{network_more}
\alias{network_more}
\title{network_more}
\usage{
network_more(output_regpcond, cytoscape = TRUE, group1 = NULL, group2 = NULL)
}
\arguments{
\item{output_regpcond}{Output object of RegulationPerCondition applied to MORE main function.}

\item{cytoscape}{TRUE for plotting the network in Cytoscape. FALSE to plot the network in R.}

\item{group1}{Name of the group to take as reference in the differential network creation.}

\item{group2}{Name of the group to compare to the reference in the differential network creation.}
}
\value{
Plot of the network induced from more.
}
\description{
\code{network_more} Function to be applied to RegulationPerConidtion function output.
}
