% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{ORA_more}
\alias{ORA_more}
\title{ORA_more}
\usage{
ORA_more(
  output,
  output_globregincond,
  annotation,
  alpha = 0.05,
  p.adjust.method = "fdr"
)
}
\arguments{
\item{output}{Output object of running more}

\item{output_globregincond}{Output object of running RegulationInCondition function}

\item{annotation}{Annotation matrix with genes in the first column, GO terms in the second}

\item{alpha}{The adjusted pvalue cutoff to consider}

\item{p.adjust.method}{One of holm, hochberg, hommel, bonferroni, BH, BY, fdr or none}
}
\value{
Plot of the network induced from more.
}
\description{
\code{ORA_more} Function to be applied to RegulationInCondition function output.
}
