% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{oraMORE}
\alias{oraMORE}
\title{oraMORE}
\usage{
oraMORE(
  output,
  outputRegincond,
  byHubs = FALSE,
  byOmic = NULL,
  annotation,
  alpha = 0.05,
  p.adjust.method = "fdr",
  parallel = FALSE
)
}
\arguments{
\item{output}{Output object of running more}

\item{outputRegincond}{Output object of running RegulationInCondition function}

\item{byHubs}{Indicates whether to perform the ORA for the Hub target features, TRUE, or for the target features regulated by the global regulators, FALSE. By default, TRUE.}

\item{byOmic}{If provided (it must follow the same nomenclature that in regulatoryData), it performs the ORA to the regulators of the specified omic. Incompatible with other methodologies specified in byHubs parameter. By default, NULL.}

\item{annotation}{Annotation matrix with target features in the first column, GO term accession in the second and GO term name in the third.}

\item{alpha}{The adjusted pvalue cutoff to consider}

\item{p.adjust.method}{One of holm, hochberg, hommel, bonferroni, BH, BY, fdr or none}

\item{parallel}{parallel If FALSE, MORE will be run sequentially. If TRUE, MORE will be run using parallelization with as many cores as the available ones minus one so the system is not overcharged. If the user wants to specify how many cores they want to use, they can also provide the number of cores to use in this parameter.}
}
\value{
Plot of the network induced from more.
}
\description{
\code{oraMORE} Function to be applied to RegulationInCondition function output.
}
