% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{oraPlot}
\alias{oraPlot}
\title{oraPlot}
\usage{
oraPlot(outputORA, regulator = NULL, Sig = c("pval", "adjPval")[1], n = 10)
}
\arguments{
\item{outputORA}{Output object of running oraMORE function}

\item{regulator}{Name of the regulator of the ORA plot the user wants to explore}

\item{Sig}{Annotation matrix with target features in the first column, pathway ID in the second and pathway name in the third. Only necessary when a specific pathway has to be plotted. By default, NULL.}

\item{n}{Number of pathways to visualize. In case there are more than 20 pathways to plot. By default, 10.}
}
\value{
ORA plot of a specific regulator
}
\description{
\code{oraPlot} Function to plot the ORA results of a specific regulator
}
