% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MORE_GLM.R
\name{plotGLM}
\alias{plotGLM}
\title{Title}
\usage{
plotGLM(GLMoutput, gene, regulator = NULL, reguValues = NULL,
  plotPerOmic = FALSE, gene.col = 1, regu.col = NULL, order = TRUE,
  xlab = "", cont.var = NULL, cond2plot = NULL, ...)
}
\arguments{
\item{order}{Should the experimental groups be ordered for the plot? If TRUE, omic values are also ordered accordingly.}

\item{...}{}
}
\description{
Title
}
