% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{plotmore}
\alias{plotmore}
\title{plotmore}
\usage{
plotmore(
  output,
  gene,
  regulator = NULL,
  simplify = FALSE,
  reguValues = NULL,
  plotPerOmic = FALSE,
  gene.col = 1,
  regu.col = NULL,
  order = TRUE,
  xlab = "",
  cont.var = NULL,
  cond2plot = NULL,
  ...
)
}
\arguments{
\item{output}{Output object of MORE main function.}

\item{gene}{ID of the gene to be plotted.}

\item{regulator}{ID of the regulator to be plotted. If NULL (default), all regulators of the gene are plotted.}

\item{reguValues}{Vector containing the values of a regulator. If NULL (default), these values are taken from the output object as long as they are available.}

\item{plotPerOmic}{If TRUE, all the relevant/significant regulators of the given gene and the same omic are plotted in the same graph. If FALSE (default), each regulator is plotted in a separate plot.}

\item{gene.col}{Color to plot the gene. By default, 1 (black).}

\item{order}{If TRUE (default), the values in X-axis are ordered.}

\item{xlab}{Label for the X-axis.}

\item{cont.var}{Vector with length equal to the number of observations in data, which optionally may contain the values of the numerical variable (e.g. time) to be plotted on the X-axis. By default, NULL.}

\item{cond2plot}{Vector or factor indicating the experimental group of each value to represent. If NULL (default), the labels are taken from the experimental design matrix.}

\item{regul.col}{Color to plot the regulator. If NULL (default), a color will be assigned by the function, that will be different for each regulatory omic.}
}
\value{
Graphical representation of the relationship between genes and regulators.
}
\description{
\code{plotmore} Graphical representation of the relationship between genes and regulators.
}
