% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{plotscores}
\alias{plotscores}
\title{plotscores}
\usage{
plotscores(output, gene, axe1 = 1, axe2 = 2)
}
\arguments{
\item{output}{Output object of MORE main function.}

\item{gene}{Gene of which the user wants visualize the weightings of the model.}

\item{axe1}{Number of the PLS component to plot in the X axis}

\item{axe2}{Number of the PLS component to plot in the Y axis}
}
\value{
Plots the scores of the samples under the PLS model generated by the significant regulators for the selected gene.
}
\description{
\code{plotscores} Function to be applied to \link{more} main function output.
}
