% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{plotweight}
\alias{plotweight}
\title{plotweight}
\usage{
plotweight(output, gene, axe1 = 1, axe2 = 2)
}
\arguments{
\item{output}{Output object of MORE main function.}

\item{gene}{Gene of which the user wants visualize the weightings of the model.}

\item{axe1}{Number of the PLS component to plot in the X axis}

\item{axe2}{Number of the PLS component to plot in the Y axis}
}
\value{
Plots the weighting star of the regulators identified as significant for the selected gene.
}
\description{
\code{plotweight} Function to be applied to \link{more} main function output.
}
