% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MORE_GLM.R
\name{RegulationPerCondition}
\alias{RegulationPerCondition}
\title{Summary table}
\usage{
RegulationPerCondition(getGLMoutput, betaTest = TRUE)
}
\arguments{
\item{getGLMoutput}{It is the name of the object where we have saved the execution of the function \code{getGLM}.}

\item{betaTest}{It is an input parameter that allows the user to perform hypothesis contrasts on the coefficients 
of the final GLM model (by default is TRUE) or not to apply any hypothesis contrast (the user will be able to 
indicate FALSE).}
}
\value{
The output is a summary table that shows for each gene the significant regulators with its omic, area, which is 
representative (R) of the correlated group and the coefficients of the final model GLM.
}
\description{
The function \code{RegulationPerCondition} returns a summary table that contains for 
each gene the regulators that are significant for each GLM model. Moreover, it shows 
the regression coefficients of the model of each regulator in the experimental covariates 
and, in the case of having correlated regulators, which regulator is the representative 
of each group of correlated regulators.
}
\examples{
data(TestData)
require(MASS); require(igraph); require(car); require(glmnet); require(psych)

## Omic type
OmicType = c(1, 0, 0)
names(OmicType) = names(data.omics)

## GetGLM function
SimGLM = GetGLM(GeneExpression = GeneExpressionDE,
                associations = associations,
                data.omics = data.omics,
                edesign = edesign[,-1, drop = FALSE],
                degree = 1,
                Res.df = 10, epsilon = 0.00001,
                alfa = 0.05, MT.adjust = "fdr",
                family = negative.binomial(theta = 10),
                elasticnet = 0.7,
                stepwise = "two.ways.backward",
                interactions.exp = TRUE, interactions.reg = 1,
                correlation = 0.8, min.variation = 0,
                cont.var = NULL, min.obs = 10, omic.type = OmicType)

myresults = RegulationPerCondition(SimGLM)


}
\seealso{
\code{\link{GetGLM}}

\code{\link{TestData}}
}
