% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{summary.MORE}
\alias{summary.MORE}
\title{summary}
\usage{
\method{summary}{MORE}(object, plot.more = FALSE)
}
\arguments{
\item{object}{MORE object obtained from applying \link{more} function.}

\item{plot.more}{If TRUE top 10 global regulators are plotted against the genes they regulate. By default, FALSE.}
}
\value{
Summary of more analysis.
}
\description{
\code{summary.MORE} Function to be applied to MORE object.
}
