% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{summary_plot}
\alias{summary_plot}
\title{summary_plot}
\usage{
summary_plot(output, output_regpcond, by_genes = TRUE)
}
\arguments{
\item{output}{MORE object obtained from applying \link{more} function.}

\item{output_regpcond}{Object generated by the function \link{RegulationPerCondition} when applied to a \link{more} object.}

\item{by_genes}{If TRUE, the function plots the percentage of genes with significant regulators globally and per omic. If FALSE, it plots the percentage of significant regulations per omic. By default, TRUE.}
}
\value{
Summary plot of the MORE analysis.
}
\description{
\code{summary_plot} Function to be applied to MORE object and the object obtained from RegulationPerCondition.
}
