% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{upsetMORE}
\alias{upsetMORE}
\title{upsetMORE}
\usage{
upsetMORE(listRegincond, byHubs = TRUE)
}
\arguments{
\item{listRegincond}{List with the output of RegulationInCondition for the conditions to be compared. It must be a named list with the condition as names.}

\item{byHubs}{Indicates whether to create the UpSet for the Hub target features, TRUE, or for the Global Regulators, FALSE. By default, TRUE.}
}
\value{
UpSet plot of the hub target features/global regulators induced from more.
}
\description{
\code{upsetMORE} Function to be applied to a list of RegulationInCondition function outputs.
}
