% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MORE_ISGL.R
\name{GetISGL}
\alias{GetISGL}
\title{\code{GetISGL} fits a MLR model with Iterative Sparse Group Lasso (ISGL) penalization
for all the features in the target omic to identify the conditions and potential
regulators that show a relevant effect on the expression of each target omic feature.}
\usage{
GetISGL(
  targetData,
  regulatoryData,
  associations = NULL,
  omicType = 0,
  condition = NULL,
  clinic = NULL,
  clinicType = NULL,
  interactions = TRUE,
  minVariation = 0,
  scaleType = "auto",
  gr.method = "MF",
  thres = 0.7
)
}
\arguments{
\item{targetData}{Data frame containing data from a target omic with its features in rows and
samples (in different conditions) in columns. Row names must be the target omic features IDs; e.g. when Gene Expression is considered gene IDs.}

\item{regulatoryData}{List where each element corresponds to a different omic data type to be considered (miRNAs,
transcription factors, methylation, etc.). The names of the list will represent the omics, and each element in 
the list should be a data matrix with omic regulators in rows and samples in columns.}

\item{associations}{List where each element corresponds to a different omic data type (miRNAs,
transcription factors, methylation, etc.). The names of the list will represent the omics. Each element in 
the list should be a data frame with 2 columns (optionally 3), describing the potential interactions between target omic features
and regulators for that omic. First column must contain the target omic features (the ones of 
\code{targetData} object), second column must contain the regulators, and an optional third column can
be added to describe the type of interaction (e.g., for methylation, if a CpG site is located in
the promoter region of the gene, in the first exon, etc.). If the user lacks prior knowledge of the potential regulators, they can set the parameter to NULL. 
In this case, all regulators in \code{regulatoryData} will be treated as potential regulators for all target omic features. In this case, for computational efficiency, it is recommended to use PLS2 \code{method}.
Additionally, if the users have prior knowledge for certain omics and want to set other omics to NULL, they can do so.}

\item{omicType}{Vector which indicates the type of data of omics introduced in \code{regulatoryData}. The user should code as 0 numeric omics and as 1 categorical or binary omics. 
By default is set to NULL. In this case, the data type will be predicted automatically. However, the user must verify the prediction and manually input the vector if incorrect.}

\item{condition}{Data frame describing the condition or phenotype to which considered samples belong. Rows must be the samples (columns
in \code{targetData}) and columns must be the conditions or phenotypes to be included in the model (e.g., treatment, etc.).}

\item{clinic}{Data frame with all clinical variables to consider, with samples in rows and variables in columns.}

\item{clinicType}{Vector which indicates the type of data of variables introduced in \code{clinic}. The user should code as 0 numeric variables and as 1 categorical or binary variables. 
By default is set to NULL. In this case, the data type will be predicted automatically. However, the user must verify the prediction and manually input the vector if incorrect.}

\item{interactions}{If TRUE, the model includes interactions between regulators and condition variables. By default, TRUE.}

\item{minVariation}{For numerical regulators, it specifies the minimum change required across conditions to retain the regulator in 
the regression models. In the case of binary regulators, if the proportion of the most common value is equal to or inferior this value, 
the regulator is considered to have low variation and will be excluded from the regression models. The user has the option to set a single 
value to apply the same filter to all omics, provide a vector of the same length as omics if they want to specify different levels for each omics, 
or use 'NA' when they want to apply a minimum variation filter but are uncertain about the threshold. By default, 0.}

\item{scaleType}{Type of scaling to be applied. Four options:
\itemize{
\item auto : Applies the auto scaling. 
\item softBlock : Applies the pareto scaling. \deqn{\frac{X_k}{s_k \sqrt[4]{m_b}} }
\item hardBlock : Applies the block scaling. \deqn{ \frac{X_k}{s_k \sqrt{m_b}} }
\item none : It does not apply any type of scaling. Not recommended if the user does not apply their own scaling.
}
considering m_b the number of variables of the block. By default, auto.}

\item{thres}{Value to determine the presence of collinearity between two regulators and consider them as group. By default, 0.7.}
}
\value{
List containing the following elements:
\itemize{
\item ResultsPerTargetF : List with as many elements as features of the target omic in \code{targetData}. For each feature, it includes information about the feature values, considered variables, estimated coefficients,
                   detailed information about all regulators, and regulators identified as relevant (in MLR scenario) or significant (in PLS scenarios).
\item GlobalSummary : List with information about the fitted models, including model metrics, information about regulators, features of the target omic without models, regulators, master regulators and hub target features.
\item Arguments : List containing all the arguments used to generate the models.                
}
}
\description{
\code{GetISGL} fits a MLR model with Iterative Sparse Group Lasso (ISGL) penalization
for all the features in the target omic to identify the conditions and potential
regulators that show a relevant effect on the expression of each target omic feature.
}
