% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{gseaMORE}
\alias{gseaMORE}
\title{gseaMORE}
\usage{
gseaMORE(
  outputRegincond,
  outputRegincond2 = NULL,
  annotation,
  alpha = 0.05,
  p.adjust.method = "fdr"
)
}
\arguments{
\item{outputRegincond}{Output object of running RegulationInCondition function}

\item{outputRegincond2}{Output object of running RegulationInCondition function for other group different to the previous. By default, NULL.}

\item{annotation}{Annotation matrix with target features in the first column, GO terms in the second and GO term description in the third}

\item{alpha}{The adjusted pvalue cutoff to consider}

\item{p.adjust.method}{One of holm, hochberg, hommel, bonferroni, BH, BY, fdr or none}
}
\value{
Plot of the network induced from more.
}
\description{
\code{gseaMORE} Function to be applied to RegulationInCondition function output.
}
