% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more.R
\name{more}
\alias{more}
\alias{more_data}
\title{Retrieve more context associated with a condition (error, warning etc.)}
\usage{
more()

more_data()
}
\description{
This function displays any detailed explanation messages and or additonal data
sets attached to a condition like an error or warning with \code{\link[=with_more]{with_more()}}.
}
\details{
Datasets are returned so can be assigned from the return of this function.

It is also possible to automatically assign datasets after a \code{more()} call
with option \code{more_auto_assign_data}, which binds them to \code{.more} in the
global environment.

\code{more()} can be called multiple times if desired. The context returned is
always for the last signalled condition that attached context using the
\code{\link[=with_more]{with_more()}}.

\code{more_data()} is a shortcut that returns the data portion of the attached
context, skipping the detailed explanation.
}
