% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{plotWeight}
\alias{plotWeight}
\title{plotWeight}
\usage{
plotWeight(output, targetF, axe1 = 1, axe2 = 2)
}
\arguments{
\item{output}{Output object of MORE main function.}

\item{targetF}{Target feature of which the user wants visualize the weightings of the model.}

\item{axe1}{Number of the PLS component to plot in the X axis}

\item{axe2}{Number of the PLS component to plot in the Y axis}
}
\value{
Plots the weighting star of the regulators identified as significant for the selected target feature.
}
\description{
\code{plotWeight} Function to be applied to \link{more} main function output.
}
