% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{RegulationInCondition}
\alias{RegulationInCondition}
\title{RegulationInCondition}
\usage{
RegulationInCondition(outputRegpcond, cond)
}
\arguments{
\item{outputRegpcond}{Output object of \link{RegulationPerCondition} function output.}

\item{cond}{Biological condition from which the user wants to summarize the information.}
}
\value{
List with the hub target features, global regulators and the regulators with their coefficients specific to the requested condition.
}
\description{
\code{RegulationInCondition} Function to be applied to \link{RegulationPerCondition} function output.
}
