% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{summaryPlot}
\alias{summaryPlot}
\title{summaryPlot}
\usage{
summaryPlot(output, outputRegpcond, byTargetF = TRUE)
}
\arguments{
\item{output}{MORE object obtained from applying \link{more} function.}

\item{outputRegpcond}{Object generated by the function \link{RegulationPerCondition} when applied to a \link{more} object.}

\item{byTargetF}{If TRUE, the function plots the percentage of target features with significant regulators globally and per omic. If FALSE, it plots the percentage of significant regulations per omic. By default, TRUE.}
}
\value{
Summary plot of the MORE analysis.
}
\description{
\code{summaryPlot} Function to be applied to MORE object and the object obtained from RegulationPerCondition.
}
